requirements.txt
fastapi
uvicorn[standard]
bilibili-api-dev
httpx
python-multipart
aiohttp
chompjs
curl_cffi
